Fix missing spdx license expression in license detection
Occassionally, the spdx license expression is missing in license detections
even though the license expression itself is non-null and a matching
spdx expression would be available. At least some instances of this
bug are due to post processing of detections, in which only the
license expression but not the spdx license expression is corrected for
license clues that are converted to license detections in post processing.

Signed-off-by: alexzurbonsen <alexander.zur.bonsen@tngtech.com>